pagina terminada david mamalo

+++ core/authentication/views.py
@@ -1,10 +1,11 @@
 # Import necessary modules and models
 from django.shortcuts import render, redirect
 from django.contrib import messages
-from django.contrib.auth import authenticate, login
+from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from .models import *
+from .forms import PasswordVerificationForm
 
 @login_required
 def home(request):
@@ -23,45 +24,49 @@ def login_page(request):
         
         if not User.objects.filter(username=username).exists():
             messages.error(request, 'Usuario no encontrado')
-            return redirect('/login/')
+            return redirect('signin')
         
         user = authenticate(username=username, password=password)
         
         if user is None:
             messages.error(request, "Contraseña incorrecta")
-            return redirect('/login/')
+            return redirect('signin')
         else:
             login(request, user)
-            return redirect('/home/')
+            return redirect('home')
     
     return render(request, 'login.html')
 
-def register_page(request):
+def signup(request):
     if request.method == 'POST':
         first_name = request.POST.get('first_name')
         last_name = request.POST.get('last_name')
         username = request.POST.get('username')
         password = request.POST.get('password')
+        confirm_password = request.POST.get('confirm_password')
         
-        user = User.objects.filter(username=username)
+        if password != confirm_password:
+            messages.error(request, "Las contraseñas no coinciden")
+            return redirect('signup')
         
-        if user.exists():
-            messages.info(request, "Usuario ya registrado")
-            return redirect('/register/')
+        if User.objects.filter(username=username).exists():
+            messages.error(request, "El nombre de usuario ya está registrado")
+            return redirect('signup')
         
         user = User.objects.create_user(
             first_name=first_name,
             last_name=last_name,
-            username=username
+            username=username,
+            password=password
         )
-        
-        user.set_password(password)
         user.save()
         
         messages.info(request, "Usuario registrado con éxito")
-        return redirect('/register/')
+        return redirect('/signin/')
     
-    return render(request, 'register.html')
+    else:
+        context = {'user_is_authenticated': request.user.is_authenticated}
+        return render(request, 'register.html', context)
 
 @login_required
 def edit_user(request, user_id):
@@ -73,12 +78,30 @@ def edit_user(request, user_id):
         user.save()
         messages.success(request, 'Usuario actualizado con éxito')
         return redirect('home')
-    return render(request, 'edit_user.html', {'user': user})
+    return render(request, 'edit_user.html', {'user': user, 'user_is_authenticated': request.user.is_authenticated})
 
+@login_required
 def delete_user(request, user_id):
     user = User.objects.get(id=user_id)
     if request.method == 'POST':
-        user.delete()
-        messages.success(request, 'Usuario eliminado con éxito')
-        return redirect('home')
-    return render(request, 'delete_user.html', {'user': user})
+        form = PasswordVerificationForm(request.user, request.POST)
+        if form.is_valid():
+            user.delete()
+            messages.success(request, 'Usuario eliminado con éxito')
+            return redirect('signin')
+        else:
+            messages.error(request, 'Contraseña incorrecta')
+            return redirect('delete_user', user_id=user_id)
+    else:
+        form = PasswordVerificationForm(request.user)
+        context = {
+            'user': user,
+            'form': form,
+            'user_is_authenticated': request.user.is_authenticated
+        }
+        return render(request, 'delete_user.html', context)
+
+@login_required
+def signout(request):
+    logout(request)
+    return redirect('signin')
